demo/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework == 3.12.0
djangorestframework-simplejwt==4.8.0
requests==2.26.0
drf-yasg